@@ -15,14 +17,19 @@
     },
     "private": true,
     "dependencies": {
+        "@mantine/hooks": "^6.0.5",
         "@next/font": "13.1.1",
-        "@nrwl/next": "15.8.9",
+        "@nx/next": "16.2.1",
+        "@tabler/icons-react": "^2.12.0",
         "class-variance-authority": "^0.4.0",
         "clsx": "^1.2.1",
-        "next": "13.1.1",
+        "lottie-colorify": "^0.8.0",
+        "lottie-react": "^2.4.0",
+        "next": "13.3.0",
         "next-themes": "^0.2.1",
         "react": "18.2.0",
         "react-dom": "18.2.0",
+        "react-intersection-observer": "^9.4.3",
         "tailwind-merge": "^1.10.0",
         "tailwindcss-animate": "^1.0.5",
         "tslib": "^2.3.0"
@@ -75,24 +77,34 @@
         "eslint-plugin-jsx-a11y": "6.7.1",
         "eslint-plugin-react": "7.32.2",
         "eslint-plugin-react-hooks": "4.6.0",
+        "eslint-plugin-storybook": "^0.6.12",
         "eslint-plugin-tailwindcss": "^3.10.1",
         "husky": "^8.0.3",
         "jest": "^29.4.1",
         "jest-environment-jsdom": "^29.4.1",
-        "nx": "15.8.9",
+        "jest-transform-stub": "^2.0.0",
+        "nx": "16.2.1",
+        "nx-cloud": "16.0.5",
         "postcss": "^8.4.21",
         "prettier": "^2.6.2",
-        "react-test-renderer": "18.2.0",
-        "storybook-addon-next": "^1.7.3",
-        "storybook-addon-swc": "^1.1.9",
-        "storybook-tailwind-dark-mode": "^1.0.15",
-        "tailwindcss": "^3.3.0",
-        "ts-jest": "^29.0.5",
+        "prettier-plugin-tailwindcss": "^0.2.7",
+        "storybook-tailwind-dark-mode": "^1.0.22",
+        "tailwindcss": "latest",
+        "ts-jest": "29.1.0",
         "ts-node": "10.9.1",
-        "typescript": "~4.9.5"
+        "typescript": "5.0.4"
     },
-    "resolutions": {
-        "@storybook/react-docgen-typescript-plugin": "npm:react-docgen-typescript-plugin@1.0.2"
-    },
-    "packageManager": "yarn@3.2.3"
+    "packageManager": "yarn@3.2.3",
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    }
 }
