@@ -1,36 +1,25 @@
-import { rootMain } from '../../.storybook/main';
-import type { StorybookConfig, Options } from '@storybook/core-common';
-import path from 'path';
+import type { StorybookConfig } from '@storybook/nextjs';
 
 const config: StorybookConfig = {
-    ...rootMain,
-    core: { ...rootMain.core, builder: 'webpack5' },
+    core: { disableTelemetry: true },
     stories: [
-        ...rootMain.stories,
         '../../libs/components/**/*.stories.@(js|jsx|ts|tsx|mdx)',
         '../../libs/pages/**/*.stories.@(js|jsx|ts|tsx|mdx)',
     ],
+    addons: ['@storybook/addon-essentials', 'storybook-tailwind-dark-mode'],
     staticDirs: ['../web/public'],
-    addons: [
-        ...(rootMain.addons || []),
-        '@nrwl/react/plugins/storybook',
-        'storybook-addon-swc',
-        'storybook-tailwind-dark-mode',
-        {
-            name: 'storybook-addon-next',
-            options: {
-                nextConfigPath: path.resolve(__dirname, '../web/next.config.js'),
-            },
+    framework: {
+        name: '@storybook/nextjs',
+        options: {
+            nextConfigPath: '../web/next.config.js',
         },
-    ],
-    webpackFinal: async (config, { configType }: Options) => {
-        if (rootMain.webpackFinal) {
-            config = await rootMain.webpackFinal(config, {
-                configType,
-            } as Options);
-        }
-        return config;
     },
+    docs: { autodocs: true },
+    features: { storyStoreV7: true, buildStoriesJson: true },
 };
 
-module.exports = config;
+export default config;
+
+// To customize your webpack configuration you can use the webpackFinal field.
+// Check https://storybook.js.org/docs/react/builders/webpack#extending-storybooks-webpack-config
+// and https://nx.dev/packages/storybook/documents/custom-builder-configs
