@@ -1,9 +1,10 @@
 import { getGreeting } from '../support/app.po';
+import { greeting } from '@config';
 
 describe('Home', () => {
     beforeEach(() => cy.visit('/'));
 
     it('should display welcome message', () => {
-        getGreeting().contains('Welcome to Home!');
+        getGreeting().contains(greeting.title);
     });
 });
