@@ -1,17 +1,14 @@
 //@ts-check
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const { withNx } = require('@nrwl/next/plugins/with-nx');
+const { withNx } = require('@nx/next/plugins/with-nx');
 
 /**
- * @type {import('@nrwl/next/plugins/with-nx').WithNxOptions}
+ * @type {import('@nx/next/plugins/with-nx').WithNxOptions}
  **/
 const nextConfig = {
     reactStrictMode: true,
     swcMinify: true,
-    experimental: {
-        fontLoaders: [{ loader: '@next/font/google', options: { subsets: ['latin'] } }],
-    },
     nx: {
         svgr: true,
     },
