@@ -15,7 +15,7 @@ class MyDocument extends Document {
                     <meta name='keywords' content={site.keywords} />
                     <meta name='description' content={site.description} />
                 </Head>
-                <body className='dark:text-dimmed-font dark:bg-dimmed-900 min-h-screen font-sans antialiased'>
+                <body className='dark:text-dimmed-font dark:bg-dimmed-900 h-screen scroll-smooth bg-white font-sans antialiased'>
                     <Main />
                     <NextScript />
                 </body>
