@@ -3,7 +3,7 @@ const { join } = require('path');
 module.exports = {
     plugins: {
         tailwindcss: {
-            config: join(__dirname, 'tailwind.config.js'),
+            config: join(__dirname, 'tailwind.config.ts'),
         },
     },
 };
