@@ -1,13 +1,14 @@
 import { cn } from '@styles';
+import { type HTML } from '@types';
 
 /* eslint-disable-next-line */
-export interface FooterProps {}
+export interface FooterProps extends HTML<'footer'> {}
 
 export const Footer: React.FC<FooterProps> = ({ ...props }) => {
     return (
-        <div className={cn([])}>
+        <footer {...props} className={cn([])}>
             <h1>Welcome to Footer!</h1>
-        </div>
+        </footer>
     );
 };
 
